pyproject.toml
[project]
name = "genai-meal-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "pydantic>=2.12.0",
    "pydantic-ai-slim[google]>=1.69.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.13.9",
    "httpx>=0.28.1",
    "pytest>=8.4.0",
    "pre-commit>=4.5.1",
]
